# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type NewsArticleSentiment =
  (typeof NewsArticleSentiment)[keyof typeof NewsArticleSentiment];

export const NewsArticleSentiment = {
  positive: "positive",
  neutral: "neutral",
  negative: "negative",
} as const;

export interface NewsArticle {
  id: string;
  title: string;
  summary: string;
  category: string;
  source: string;
  publishedAt: string;
  imageUrl?: string;
  /** Importance score 1-10 */
  importance: number;
  sentiment: NewsArticleSentiment;
  tags: string[];
  /** Estimated read time in seconds */
  readTime?: number;
  relevanceReason?: string;
}

export type TrendingTopicTrend =
  (typeof TrendingTopicTrend)[keyof typeof TrendingTopicTrend];

export const TrendingTopicTrend = {
  rising: "rising",
  stable: "stable",
  falling: "falling",
} as const;

export interface TrendingTopic {
  id: string;
  name: string;
  category: string;
  /** Size/importance 1-100 */
  importance: number;
  articleCount: number;
  trend: TrendingTopicTrend;
  color?: string;
}

export type SmartAlertType =
  (typeof SmartAlertType)[keyof typeof SmartAlertType];

export const SmartAlertType = {
  weather: "weather",
  traffic: "traffic",
  breaking: "breaking",
  local: "local",
  safety: "safety",
} as const;

export type SmartAlertSeverity =
  (typeof SmartAlertSeverity)[keyof typeof SmartAlertSeverity];

export const SmartAlertSeverity = {
  low: "low",
  medium: "medium",
  high: "high",
  critical: "critical",
} as const;

export interface SmartAlert {
  id: string;
  type: SmartAlertType;
  title: string;
  message: string;
  impact?: string;
  severity: SmartAlertSeverity;
  location?: string;
  timestamp: string;
}

export type BriefingRequestMode =
  (typeof BriefingRequestMode)[keyof typeof BriefingRequestMode];

export const BriefingRequestMode = {
  beginner: "beginner",
  normal: "normal",
  expert: "expert",
} as const;

export type BriefingRequestStyle =
  (typeof BriefingRequestStyle)[keyof typeof BriefingRequestStyle];

export const BriefingRequestStyle = {
  tldr: "tldr",
  deep_dive: "deep_dive",
  thirty_seconds: "thirty_seconds",
} as const;

export interface BriefingRequest {
  mode: BriefingRequestMode;
  style?: BriefingRequestStyle;
}

export interface ArticleBriefing {
  articleId: string;
  tldr: string;
  keyPoints: string[];
  whyItMatters: string;
  impact: string;
  whatNext: string;
  mode: string;
}

export type TimelineEventImportance =
  (typeof TimelineEventImportance)[keyof typeof TimelineEventImportance];

export const TimelineEventImportance = {
  low: "low",
  medium: "medium",
  high: "high",
} as const;

export interface TimelineEvent {
  date: string;
  title: string;
  description: string;
  importance: TimelineEventImportance;
}

export interface KeyPerson {
  name: string;
  role: string;
  sentiment: string;
}

export interface SentimentChange {
  date: string;
  sentiment: number;
  label: string;
}

export interface StoryTimeline {
  articleId: string;
  title: string;
  events: TimelineEvent[];
  keyPeople: KeyPerson[];
  sentimentChanges: SentimentChange[];
  futurePredictions: string[];
}

export type UserPreferencesReadingLevel =
  (typeof UserPreferencesReadingLevel)[keyof typeof UserPreferencesReadingLevel];

export const UserPreferencesReadingLevel = {
  beginner: "beginner",
  normal: "normal",
  expert: "expert",
} as const;

export interface UserPreferences {
  categories: string[];
  location?: string;
  readingLevel: UserPreferencesReadingLevel;
  alertsEnabled?: boolean;
}

export interface OpenaiConversation {
  id: number;
  title: string;
  createdAt: string;
}

export interface OpenaiMessage {
  id: number;
  conversationId: number;
  role: string;
  content: string;
  createdAt: string;
}

export interface CreateOpenaiConversationBody {
  title: string;
}

export interface SendOpenaiMessageBody {
  content: string;
}

export interface OpenaiConversationWithMessages {
  id: number;
  title: string;
  createdAt: string;
  messages: OpenaiMessage[];
}

export type GenerateOpenaiImageBodySize =
  (typeof GenerateOpenaiImageBodySize)[keyof typeof GenerateOpenaiImageBodySize];

export const GenerateOpenaiImageBodySize = {
  "1024x1024": "1024x1024",
  "512x512": "512x512",
  "256x256": "256x256",
} as const;

export interface GenerateOpenaiImageBody {
  prompt: string;
  size?: GenerateOpenaiImageBodySize;
}

export interface GenerateOpenaiImageResponse {
  b64_json: string;
}

export interface OpenaiError {
  error: string;
}

export type GetNewsFeedParams = {
  categories?: string;
  limit?: number;
  offset?: number;
};

export type GetAlertsParams = {
  location?: string;
};