# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /news/feed:
    get:
      operationId: getNewsFeed
      tags: [news]
      summary: Get personalized news feed
      parameters:
        - name: categories
          in: query
          required: false
          schema:
            type: string
        - name: limit
          in: query
          required: false
          schema:
            type: integer
        - name: offset
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: News feed
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/NewsArticle"
...
  /news/alerts:
    get:
      operationId: getAlerts
      tags: [news]
      summary: Get smart alerts
      parameters:
        - name: location
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: Smart alerts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SmartAlert"
  /news/articles/{id}/briefing:
    post:
      operationId: generateBriefing
      tags: [news]
      summary: Generate AI briefing for an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BriefingRequest"
      responses:
        "200":
          description: AI briefing
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleBriefing"
  /news/articles/{id}/timeline:
    get:
      operationId: getArticleTimeline
      tags: [news]
      summary: Get story arc timeline for an article
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Story timeline
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StoryTimeline"
  /news/preferences:
    get:
      operationId: getUserPreferences
      tags: [news]
      summary: Get user preferences
      responses:
        "200":
          description: User preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
    post:
      operationId: saveUserPreferences
      tags: [news]
      summary: Save user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPreferences"
      responses:
        "200":
          description: Saved preferences
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserPreferences"
...
components:
  schemas:
...
    NewsArticle:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        summary:
          type: string
        category:
          type: string
        source:
          type: string
        publishedAt:
          type: string
          format: date-time
        imageUrl:
          type: string
        importance:
          type: integer
          description: Importance score 1-10
        sentiment:
          type: string
          enum: [positive, neutral, negative]
        tags:
          type: array
          items:
            type: string
        readTime:
          type: integer
          description: Estimated read time in seconds
        relevanceReason:
          type: string
      required:
        - id
        - title
        - summary
        - category
        - source
        - publishedAt
        - importance
        - sentiment
        - tags
...
    SmartAlert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [weather, traffic, breaking, local, safety]
        title:
          type: string
        message:
          type: string
        impact:
          type: string
        severity:
          type: string
          enum: [low, medium, high, critical]
        location:
          type: string
        timestamp:
          type: string
          format: date-time
      required:
        - id
        - type
        - title
        - message
        - severity
        - timestamp
    BriefingRequest:
      type: object
      properties:
        mode:
          type: string
          enum: [beginner, normal, expert]
        style:
          type: string
          enum: [tldr, deep_dive, thirty_seconds]
      required:
        - mode
    ArticleBriefing:
      type: object
      properties:
        articleId:
          type: string
        tldr:
          type: string
        keyPoints:
          type: array
          items:
            type: string
        whyItMatters:
          type: string
        impact:
          type: string
        whatNext:
          type: string
        mode:
          type: string
      required:
        - articleId
        - tldr
        - keyPoints
        - whyItMatters
        - impact
        - whatNext
        - mode
    StoryTimeline:
      type: object
      properties:
        articleId:
          type: string
        title:
          type: string
        events:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"
        keyPeople:
          type: array
          items:
            $ref: "#/components/schemas/KeyPerson"
        sentimentChanges:
          type: array
          items:
            $ref: "#/components/schemas/SentimentChange"
        futurePredictions:
          type: array
          items:
            type: string
      required:
        - articleId
        - title
        - events
        - keyPeople
        - sentimentChanges
        - futurePredictions
    TimelineEvent:
      type: object
      properties:
        date:
          type: string
        title:
          type: string
        description:
          type: string
        importance:
          type: string
          enum: [low, medium, high]
      required:
        - date
        - title
        - description
        - importance
    KeyPerson:
      type: object
      properties:
        name:
          type: string
        role:
          type: string
        sentiment:
          type: string
      required:
        - name
        - role
        - sentiment
    SentimentChange:
      type: object
      properties:
        date:
          type: string
        sentiment:
          type: number
        label:
          type: string
      required:
        - date
        - sentiment
        - label
    UserPreferences:
      type: object
      properties:
        categories:
          type: array
          items:
            type: string
        location:
          type: string
        readingLevel:
          type: string
          enum: [beginner, normal, expert]
        alertsEnabled:
          type: boolean
      required:
        - categories
        - readingLevel
...