# lib/api-zod/src/generated/types/articleBriefing.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ArticleBriefing {
  articleId: string;
  tldr: string;
  keyPoints: string[];
  whyItMatters: string;
  impact: string;
  whatNext: string;
  mode: string;
}