# lib/api-zod/src/generated/types/briefingRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { BriefingRequestMode } from "./briefingRequestMode";
import type { BriefingRequestStyle } from "./briefingRequestStyle";

export interface BriefingRequest {
  mode: BriefingRequestMode;
  style?: BriefingRequestStyle;
}