# lib/api-zod/src/generated/types/newsArticle.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { NewsArticleSentiment } from "./newsArticleSentiment";

export interface NewsArticle {
  id: string;
  title: string;
  summary: string;
  category: string;
  source: string;
  publishedAt: Date;
  imageUrl?: string;
  /** Importance score 1-10 */
  importance: number;
  sentiment: NewsArticleSentiment;
  tags: string[];
  /** Estimated read time in seconds */
  readTime?: number;
  relevanceReason?: string;
}